abstracted out model reduction to cleaned up the code
